+++ src/main/java/roomescape/config/JwtProvider.java
@@ -8,9 +8,7 @@ import io.jsonwebtoken.SignatureAlgorithm;
 import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.domain.MemberRepository;
 import roomescape.exception.AuthorizationException;
-import roomescape.service.request.TokenAppRequest;
 
 @Component
 public class JwtProvider {
@@ -20,26 +18,17 @@ public class JwtProvider {
     @Value("${security.jwt.token.expire-length}")
     private long validityInMilliseconds;
 
-    private final MemberRepository memberRepository;
+    public String createToken(String email) {
+        Claims claims = Jwts.claims().setSubject(email);
+        Date now = new Date();
+        Date validity = new Date(now.getTime() + validityInMilliseconds);
 
-    public JwtProvider(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
-
-    public String createToken(TokenAppRequest request) {
-        if (isExistsMember(request.email(), request.password())) {
-            Claims claims = Jwts.claims().setSubject(request.email());
-            Date now = new Date();
-            Date validity = new Date(now.getTime() + validityInMilliseconds);
-
-            return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(now)
-                .setExpiration(validity)
-                .signWith(SignatureAlgorithm.HS256, secretKey)
-                .compact();
-        }
-        throw new AuthorizationException("이메일 또는 비밀번호가 잘못되었습니다.");
+        return Jwts.builder()
+            .setClaims(claims)
+            .setIssuedAt(now)
+            .setExpiration(validity)
+            .signWith(SignatureAlgorithm.HS256, secretKey)
+            .compact();
     }
 
     public String createExpiredToken(String token) {
@@ -64,10 +53,6 @@ public class JwtProvider {
         throw new AuthorizationException("토큰이 만료되었습니다. 다시 로그인해주세요.");
     }
 
-    private boolean isExistsMember(String email, String password) {
-        return memberRepository.isExistsByEmailAndPassword(email, password);
-    }
-
     private boolean validateToken(String token) {
         try {
             Jws<Claims> claims = Jwts.parser()
@@ -80,7 +65,7 @@ public class JwtProvider {
         }
     }
 
-    private Claims getClaims(final String token) {
+    private Claims getClaims(String token) {
         return Jwts.parser()
             .setSigningKey(secretKey)
             .parseClaimsJws(token)
