--- src/main/java/roomescape/service/request/TokenAppRequest.java
@@ -1,4 +1,0 @@
-package roomescape.service.request;
-
-public record TokenAppRequest(String email, String password) {
-}
