+++ src/test/java/roomescape/controller/ControllerTest.java
@@ -1,9 +1,7 @@
 package roomescape.controller;
 
 import static roomescape.Fixture.VALID_ADMIN_EMAIL;
-import static roomescape.Fixture.VALID_ADMIN_PASSWORD;
 import static roomescape.Fixture.VALID_USER_EMAIL;
-import static roomescape.Fixture.VALID_USER_PASSWORD;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
@@ -12,8 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.service.JwtProvider;
-import roomescape.service.request.TokenAppRequest;
+import roomescape.config.JwtProvider;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql(scripts = "/truncate.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
@@ -32,12 +29,10 @@ public class ControllerTest {
     }
 
     String getUserToken() {
-        return jwtProvider.createToken(
-            new TokenAppRequest(VALID_USER_EMAIL.getValue(), VALID_USER_PASSWORD.getValue()));
+        return jwtProvider.createToken(VALID_USER_EMAIL.getValue());
     }
 
     String getAdminToken() {
-        return jwtProvider.createToken(
-            new TokenAppRequest(VALID_ADMIN_EMAIL.getValue(), VALID_ADMIN_PASSWORD.getValue()));
+        return jwtProvider.createToken(VALID_ADMIN_EMAIL.getValue());
     }
 }
